Get rid of MatchableResource interface, since it is internal and there is only one implementation
